stt/server/requirements.txt
# NeMo STT Server dependencies
# Install in WSL conda environment: conda activate nemo && pip install -r requirements.txt

nemo_toolkit[asr]
numpy
